rename eigen_types.h to .hpp since the cmakefile only copy .hpp files from ./src/ into installation folder...
