+++ orocos_kdl/src/articulatedbodyinertia.cpp
@@ -20,7 +20,7 @@
 // Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
 
 #include "articulatedbodyinertia.hpp"
-#include "eigen_types.h"
+#include "eigen_types.hpp"
 
 namespace KDL{
