+++ orocos_kdl/src/chainiksolverpos_lma.hpp
@@ -33,7 +33,7 @@
 
 #include "chainiksolver.hpp"
 #include "chain.hpp"
-#include "eigen_types.h"
+#include "eigen_types.hpp"
 
 namespace KDL
 {
