+++ orocos_kdl/src/chainiksolvervel_pinv_givens.hpp
@@ -5,7 +5,7 @@
 
 #include "chainiksolver.hpp"
 #include "chainjnttojacsolver.hpp"
-#include "eigen_types.h"
+#include "eigen_types.hpp"
 
 namespace KDL
 {
