+++ orocos_kdl/src/chainiksolvervel_wdls.hpp
@@ -24,7 +24,7 @@
 
 #include "chainiksolver.hpp"
 #include "chainjnttojacsolver.hpp"
-#include "eigen_types.h"
+#include "eigen_types.hpp"
 
 
 namespace KDL
